+++ CHANGELOG.md
@@ -6,6 +6,10 @@ All notable changes to this project will be documented in this file.
  - Boards *createPrivate*, *sendWithMessage* and *sendWithEmail* methods.
  - Pins *sendWithMessage* and *sendWithEmail* methods.
 
+### Changed:
+ - Removed $removeCookies param from *Auth::logout()* method. Use 
+ *bot->getHttpClient()->removeCookies()* to remove your cookies.
+
 ## [5.0.1] - 2016-11-13
 ### Fixed:
  - Interests *main* method requires to be logged in.